+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). Currently, this project is using an `0.x` versioning system to indicate its preview status. As such, it does _not_ currently adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). Once we reach production readiness, we will
 follow semver beginning with our `1.0` release. In the meantime, please read this changelog for information on breaking changes!
 
+## 0.21.0 [March 12, 2019]
+
+### Added
+
+- Select: add `getOptionValue` support. [#1555]
+
+### Fixed
+
+- Do not try to clone null elements. [#1554]
+
 ## 0.20.3 [March 11, 2019]
 
 ### Fixed